TY-2250 disable some lints in generated files [2] (#4)
* TY-2250 unreachable_pub warning

No mangle functions are implicit pub, so there is no point
in not making them rust-pub. But they might not be exported
to the normal rust library api, in which case this lint would
trigger, so the lint got disabled for the generated code.

* allow(clippy::module_name_repetitions) in generated files

We need this to avoid name collisions of non-mangled names.